pyproject.toml
[tool.poetry]
name = "anole"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = "^2.4.1"
accelerate = "^1.0.1"
deepspeed = "^0.15.2"
jsonlines = "^4.0.0"
pandas = "^2.2.3"
pillow = "^10.4.0"
pyarrow = "^17.0.0"
numpy = "^2.1.2"
python-dotenv = "^1.0.1"
click = "^8.1.7"
flask = "^3.0.3"
flask-socketio = "^5.4.1"
xformers = "^0.0.28.post1"
asyncio = "^3.4.3"
aiofiles = "^24.1.0"
tokenizers = "^0.20.1"
transformers = "^4.45.2"
regex = "^2024.9.11"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
black = "^24.10.0"
flake8 = "^7.1.1"
matplotlib = "^3.9.2"
types-aiofiles = "^24.1.0.20240626"
pandas-stubs = "^2.2.3.241009"
types-tqdm = "^4.66.0.20240417"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
